transaction_simulator/requirements.txt
azure-core==1.30.0
azure-eventhub==5.11.5
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
numpy==1.26.4
pandas==2.2.0
python-dateutil==2.8.2
pytz==2024.1
random2==1.0.2
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
